# Improving performance of `cupy.random.choice`