Fix https://github.com/cupy/cupy/issues/8477.

current master:

```
CPU:   528.909 us   +/-  8.817 (min:   513.830 / max:   556.690) us     GPU-0: 22175.946 us   +/- 21.370 (min: 22132.736 / max: 22215.679) us
```

This PR:

```
CPU:   542.260 us   +/- 226.564 (min:   510.556 / max:  2789.806) us     GPU-0:   548.843 us   +/- 227.403 (min:   517.088 / max:  2804.736) u
```

Reproducer:

```py
import cupy as cp
from cupyx.profiler import benchmark

def f():
    cp.random.choice(10000, size=10000, p=cp.ones(10000)/10000)

print(benchmark(f, (), n_repeat=20))
```