@seberg 2024-08-19
This PR changes the stream of random numbers (noticed in a cuML doctest, I suspect that isn't a huge deal, once can set up the doctests to only run on newer cupy presumably).

I am not sure if CuPy has much of a policy around stream stability or if you may want to document the change.

Old stream:
```
In [5]: cp.random.RandomState(cp.uint32(10)).choice([0, 1], size=20, p=[0.5, 0.5])
Out[5]: array([1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0])
```
New stream:
```
In [2]: cp.random.RandomState(cp.uint32(10)).choice([0, 1], size=20, p=[0.5, 0.5])
Out[2]: array([0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1])
```