@jakirkham 2024-08-20
Curious what the result is if we drop `p` (given there is equal probability of either option)?