@kmaehashi 2024-08-20
Thanks for the catch! There's no explicit policy, but I think in general users would expect reproducibility between minor versions with fixed seeds, so I'll make a revert PR for v13 backport. cc/ @asi1024