@jakirkham 2024-08-20
Sounds good. Thanks all! 🙏

Given this code is more like NumPy (and it is more performant), think that is what we want longer term. We probably just need a little more time to understand the differences between the two

So reverting in v13 seems like the right approach